% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Conjunto de datos: Datos}
\format{
conjunto de datos con tantas filas y tantas columnas
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por cinco}
}
}
\source{
Repositorio de Github.
}
\usage{
datos
}
\description{
descripcion
}
\keyword{datasets}
