% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primera_funcion.R
\name{descarga_datos}
\alias{descarga_datos}
\title{Funcion de descarga de datos}
\usage{
descarga_datos(id_estacion, ruta_archivo)
}
\arguments{
\item{ruta_archivo}{}
}
\value{
un archivo .csv con el id de la estación buscada
}
\description{
Funcion de descarga de datos
}
\examples{
descarga_datos("NH0098", "C:/Users/Documentos/windyR/datos/estacion_NH0098")
descarga_datos("NH0910", "C:/Users/Documentos/windyR/datos/estacion_NH0910")
}
