% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura.R
\name{grafico_mensual}
\alias{grafico_mensual}
\title{Grafico de temperatura mensual}
\usage{
grafico_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\value{
Nos devuelve un grafico con la estacion seleccionada, con el eje x sobre el mes y el eje y sobre temperatura promedio
}
\description{
Grafico de temperatura mensual
}
\examples{
grafico_temperatura_mensual(estacion_NH0046, colores = "red", titulo = "Temperatura")
}
