% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_temperatura.R
\name{tabla_de_temperatura_estaciones}
\alias{tabla_de_temperatura_estaciones}
\title{Tabla de temperatura de estaciones}
\usage{
tabla_de_temperatura_estaciones(datos)
}
\value{
Nos devuelve una tabla con valores de temperatura
}
\description{
Tabla de temperatura de estaciones
}
\examples{
tabla_de_temperatura_estaciones(estacion_NH0046)

}
