% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primera_funcion.R
\name{descarga_datos}
\alias{descarga_datos}
\title{Funcion de descarga de datos}
\usage{
descarga_datos(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{Un string con el ID de la estación meteorológica (por ejemplo, "NH0046"). Este ID se utiliza para construir la URL de descarga del archivo.}

\item{ruta_archivo}{Un string con la ruta completa y el nombre del archivo donde se guardará el CSV descargado.}
}
\value{
Un data frame con los datos de la estación meteorológica descargada.
}
\description{
La funcion \code{descarga_datos()} descarga un archivo CSV de datos de una estación meteorológica específica desde un repositorio en línea y lo guarda en la ruta especificada. Luego, lee el archivo y devuelve los datos como un data frame.
}
\examples{
\dontrun{
descarga_datos("NH0098", "C:/inserte/su/ruta/de/directorio/datos/estacion_NH0098")
}

}
