% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primera_funcion.R
\name{descarga_datos}
\alias{descarga_datos}
\title{Funcion de descarga de datos}
\usage{
descarga_datos(id_estacion, ruta_archivo)
}
\arguments{
\item{ruta_archivo}{}
}
\description{
Funcion de descarga de datos
}
