% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura.R
\name{grafico_mensual}
\alias{grafico_mensual}
\title{Title}
\usage{
grafico_mensual(datos, paleta = NULL, titulo_grafico = "Temperatura Mensual")
}
\arguments{
\item{titulo_grafico}{}
}
\description{
Title
}
